bug: 순환참조 디버깅 (#404)
* refactor: 챌린지 관련 모달 위치 변경 (shared -> features)

* feat: 참여 챌린지 페이지 Suspense 감싸기

* feat: FSD eslint 재설정

* refactor: FSD 파일/폴더 위치 재조정(챗봇)

* refactor: 챌린지 알람 리스트 FSD 레이어 규칙 적용

* chore: eslint 설정 끄기
