+++ src/entities/challenge/api/alarm/get-alarm-list.ts
@@ -1,32 +1,32 @@
-import { ChallengeType } from '@/entities/challenge/model'
-
 import { ENDPOINTS, fetchRequest } from '@/shared/lib'
 import { InfiniteScrollResponse, ISOFormatString } from '@/shared/type'
 
-export interface MemberAlarmListParams {
+import { ChallengeType } from '../../model'
+
+export interface ChallengeAlarmListParams {
   cursorId?: number
   cursorTimestamp?: ISOFormatString
 }
 
-export type AlarmType = {
+export type AlarmType<T> = {
   id: number // 알람 아이디
   title: string
   content: string
   createdAt: ISOFormatString
   isRead: boolean
-  type: ChallengeType
+  type: T
   imageUrl: string
   challengeId: number
 }
 
-export type MemberAlarmList = InfiniteScrollResponse<{
-  notifications: AlarmType[]
+export type ChallengeAlarmType = AlarmType<ChallengeType>
+
+export type ChallengeAlarmList = InfiniteScrollResponse<{
+  notifications: ChallengeAlarmType[]
 }>
 
-type GetMemberAlarmListResponse = MemberAlarmList
-
-export const getMemberAlarmList = ({ cursorId, cursorTimestamp }: MemberAlarmListParams) => {
-  return fetchRequest<GetMemberAlarmListResponse>(ENDPOINTS.MEMBERS.NOTIFICATION.LIST, {
+export const getChallengeAlarmList = ({ cursorId, cursorTimestamp }: ChallengeAlarmListParams) => {
+  return fetchRequest<ChallengeAlarmList>(ENDPOINTS.MEMBERS.NOTIFICATION.LIST, {
     query: {
       ...(cursorId !== undefined ? { cursorId } : {}),
       ...(cursorTimestamp ? { cursorTimestamp } : {}),
