+++ src/entities/challenge/api/index.ts
@@ -1,4 +1,5 @@
 // 폴더
+export * from './alarm'
 export * from './event'
 export * from './group'
 export * from './personal'
