+++ src/entities/chatbot/model/type.ts
@@ -1,7 +1,3 @@
-import type React from 'react'
-
-import { ChatSelectionProps } from '@/features/chatbot/components/chat-selection'
-
 export type ChatOption = {
   label: string
   value: string
@@ -11,16 +7,4 @@ export type ChatSelections = {
   location?: string
   workType?: string
   category?: string
-}
-
-export type ChatHistoryItem = {
-  type: 'message' | 'selection' | 'horizontal-cards'
-  role?: 'bot' | 'user'
-  text?: React.ReactNode
-  loading?: boolean
-  selectionProps?: ChatSelectionProps
-  subDescription?: string
-  buttonText?: string
-  isAnswer?: boolean
-  onClick?: () => void
 }
