+++ src/features/challenge/components/index.ts
@@ -7,6 +7,7 @@ export * from './feed-list'
 export * from './form-detail-step'
 export * from './form-metadata-step'
 export * from './group-challenge-card'
+export * from './modal'
 export * from './verification-card'
 export * from './verification-carousel'
 export * from './verification-comment-item'
