+++ src/features/challenge/components/modal/image-zoom-modal/index.ts
