+++ src/features/challenge/components/modal/index.ts
@@ -1,0 +1,3 @@
+export * from './camera-modal'
+export * from './image-zoom-modal'
+export * from './verification-guide-modal'
