+++ src/features/challenge/components/modal/verification-guide-modal/index.ts
