+++ src/features/challenge/components/modal/verification-guide-modal/styles.ts
