+++ src/features/challenge/components/modal/verification-guide-modal/verification-guide-modal.tsx
